backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Deprecated==1.2.14
Django==4.2.3
django-filter==23.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jwcrypto==1.5.0
Markdown==3.4.3
oauthlib==3.2.2
packaging==23.1
psycopg2==2.9.6
pycparser==2.21
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
wrapt==1.15.0
